Reintroduce IAM instance profile ARN for Batch instance role
Reintroduce changes made in 1caf7fd66ef2a675769e398cc246d7790068e129
that use a data source to retrieve an IAM instance profile based on
the container_instance_ec2 IAM role. This is necessary because Batch
expects to receive the ARN of an instance profile, not the ARN of the
role contained within the instance profile envelope. Since the IAM role
is not reliably created before the data source is evaluated, add an
explicit dependency with depends_on.