Remove AMI creation and use default AMIs
This removes functionality around creating AMIs and specifying them in the
CloudFormation template. AWS Batch is now smart enough to default to the ECS-optimized
AMI for CPU compute environments and the ECS-optimized AMI with GPU support for GPU
compute environments. To make this work, we had set the ResourceRequirments field
in the job defs.